# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "This is the API documentation for my Laravel application.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/sender_customers": {
            "get": {
                "tags": [
                    "Sender Customer"
                ],
                "summary": "Retrieve a list of all sender customers",
                "operationId": "a3187d98ac3c4211eb5930a8efaa1106",
                "responses": {
                    "200": {
                        "description": "Sender customer list retrieved successfully"
                    },
                    "404": {
                        "description": "Sender customer list not found"
                    }
                }
            }
        },
        "/api/v1/sender_customer_register": {
            "post": {
                "tags": [
                    "Sender Customer"
                ],
                "summary": "Store a new sender customer",
                "operationId": "ba14955f35215a3ad65080757a9371fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sender customer created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/sender_customer_update/{id}": {
            "put": {
                "tags": [
                    "Sender Customer"
                ],
                "summary": "Update a sender customer",
                "operationId": "0b7b43ca948f2fcd918c664fc6b67631",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sender customer to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sender customer updated successfully"
                    },
                    "404": {
                        "description": "Sender customer not found"
                    }
                }
            }
        },
        "/api/v1/sender_customer_destroy/{id}": {
            "delete": {
                "tags": [
                    "Sender Customer"
                ],
                "summary": "Delete a sender customer",
                "operationId": "3ef38e2c0fa4f9d96d5a4ebce9ab2ba6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sender customer to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sender customer deleted successfully"
                    },
                    "404": {
                        "description": "Sender customer not found"
                    }
                }
            }
        },
        "/api/v1/sender_customer_show/{id}": {
            "get": {
                "tags": [
                    "Sender Customer"
                ],
                "summary": "Retrieve a specific sender customer",
                "operationId": "cd465eb65d0f9c9672a7cb75a6810a03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sender customer to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sender customer retrieved successfully"
                    },
                    "404": {
                        "description": "Sender customer not found"
                    }
                }
            }
        },
...
    },
...
}